Add conditional for parsing SERIES
The CSV from Bank of Canada is structured wherein single column titles
in the CSV will indicate how the next few lines of the CSV will be
structured

As such, detect when the `SERIES` title is encountered, and write logic
to parse the contents of the `SERIES` data into a couple of `dicts`.
One which maps the three letter code of a currency to their
corresponding `id`, and another that maps the `id` of each row to their
corresponding `label` and `description`

This will allow functions to be implemented in the future to be given a
human readable currency three letter code (i.e. 'USD', 'AUD'), and
retrieve the corresponding complex `id` Bank of Canada uses to identify
conversion of a currency to CAD